# sync: optimize for faster processing (fixes #5834)