… app/src/main/java/org/ole/planet/myplanet/utilities/SyncTimeLogger.kt