```
=== SYNC TIME SUMMARY ===
Total sync time: 7 min 38 sec (7m 38.117s)
Individual process times:
library_sync                  : 4m 47.150s ( 63%)
resource_sync                 : 2m 50.909s ( 37%)
admin_sync                    :       14ms (  0%)
courses_sync                  :        1ms (  0%)
exams_sync                    :        1ms (  0%)
team_activities_sync          :        1ms (  0%)
submissions_sync              :        1ms (  0%)
courses_progress_sync         :        1ms (  0%)
certifications_sync           :        1ms (  0%)
tablet_users_sync             :        1ms (  0%)
achievements_sync             :        1ms (  0%)
ratings_sync                  :        1ms (  0%)
news_sync                     :        1ms (  0%)
teams_sync                    :        0ms (  0%)
login_activities_sync         :        0ms (  0%)
health_sync                   :        0ms (  0%)
feedback_sync                 :        0ms (  0%)
meetups_sync                  :        0ms (  0%)
tags_sync                     :        0ms (  0%)
tasks_sync                    :        0ms (  0%)
chat_history_sync             :        0ms (  0%)
on_synced                     :        0ms (  0%)
=========================
```
in planet learning above was the full sync time before optimization can be tested in this branch https://github.com/open-learning-exchange/myplanet/tree/syncTimeLogger and below is after optimization on this branch sync time reduced by 4 minutes
```
=== SYNC TIME SUMMARY ===
Total sync time: 3 min 29 sec (3m 29.14s)
Individual process times:
library_sync                  : 1m 47.115s ( 51%)
resource_sync                 : 1m 41.845s ( 49%)
admin_sync                    :       26ms (  0%)
tags_sync                     :        2ms (  0%)
on_synced                     :        2ms (  0%)
courses_sync                  :        1ms (  0%)
exams_sync                    :        1ms (  0%)
courses_progress_sync         :        1ms (  0%)
certifications_sync           :        1ms (  0%)
meetups_sync                  :        1ms (  0%)
ratings_sync                  :        1ms (  0%)
teams_sync                    :        0ms (  0%)
login_activities_sync         :        0ms (  0%)
team_activities_sync          :        0ms (  0%)
health_sync                   :        0ms (  0%)
submissions_sync              :        0ms (  0%)
tablet_users_sync             :        0ms (  0%)
feedback_sync                 :        0ms (  0%)
achievements_sync             :        0ms (  0%)
tasks_sync                    :        0ms (  0%)
chat_history_sync             :        0ms (  0%)
news_sync                     :        0ms (  0%)
=========================
```
in planet sanpablo the resources still take long to process since it's quite alot of data to process
```
=== SYNC TIME SUMMARY ===
Total sync time: 50 min 40 sec (50m 40.652s)
Individual process times:
resource_sync                 : 40m 53.143s ( 81%)
library_sync                  : 9m 47.467s ( 19%)
on_synced                     :        9ms (  0%)
admin_sync                    :        6ms (  0%)
submissions_sync              :        2ms (  0%)
courses_progress_sync         :        2ms (  0%)
teams_sync                    :        1ms (  0%)
feedback_sync                 :        1ms (  0%)
achievements_sync             :        1ms (  0%)
ratings_sync                  :        1ms (  0%)
news_sync                     :        1ms (  0%)
courses_sync                  :        0ms (  0%)
exams_sync                    :        0ms (  0%)
login_activities_sync         :        0ms (  0%)
team_activities_sync          :        0ms (  0%)
health_sync                   :        0ms (  0%)
certifications_sync           :        0ms (  0%)
tablet_users_sync             :        0ms (  0%)
meetups_sync                  :        0ms (  0%)
tags_sync                     :        0ms (  0%)
tasks_sync                    :        0ms (  0%)
chat_history_sync             :        0ms (  0%)
=========================
```